+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Released at:
 
 2. Upgraded to mruby 3.0.0
 
+3. Compiled binary for OSX build with MacOSX11.3 SDK (Darwin19)
+
+4. Added binary for `arm64-apple-darwin19` target
+
 [Full Changelog](https://github.com/appplant/ski/compare/1.5.1...HEAD)
 
 <details><summary>Releasenotes</summary>